% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logo.R
\name{add_logo}
\alias{add_logo}
\title{add_logo}
\usage{
add_logo(grafico = last_plot())
}
\arguments{
\item{grafico}{Grafico gerado pelo ggplot. Por padrao, usa o ultimo grafico gerado}
}
\value{
Um grafico interativo em html (plotly)
}
\description{
Adiciona logo do cecon aos graficos
}
\author{
Arthur Welle
}
