% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajuste_xts.R
\name{ajuste_xts}
\alias{ajuste_xts}
\title{ajuste_xts}
\usage{
ajuste_xts(dados, col_data = 1, col_dados = 2, remover_NA = FALSE)
}
\arguments{
\item{dados}{Série a ser tratada}

\item{col_data}{Coluna da série que possui as datas. Por padrão, usa-se a coluna 1}

\item{col_dados}{Coluna(s) dos dados. Por padrão, seleciona-se a segunda coluna}
}
\value{
Série em xts
}
\description{
Transforma os dados em xts para utilizar nas demais séries do pacote
}
