% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ipeadata.R
\name{clean_ipeadata}
\alias{clean_ipeadata}
\alias{clean_ipea}
\title{Clean IPEA}
\usage{
clean_ipeadata(dado_ipea)
}
\arguments{
\item{dado_ipea}{Dados baixados do IPEADATA}
}
\value{
Um objeto xts
}
\description{
Trata os dados baixados do IPEADATA para serem utilizados com os outros pacotes da serie
}
\author{
Gabriel Petrini
}
