% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportar_dados.R
\name{exportar_dados}
\alias{exportar_dados}
\title{exportar_dados}
\usage{
exportar_dados(dados, pasta = "../Tratados", formato = ".RData")
}
\arguments{
\item{dados}{Objeto a ser exportado}

\item{pasta}{Caminho para a exportacao}

\item{formato}{Formato do arquivo final}
}
\value{
Nao modifica o objeto, apenas exporta
}
\description{
Exporta os dados no formato .RData para uma pasta
}
\author{
Gabriel Petrini
}
