% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FazTexto.AccMes.R
\name{FazTexto.AccMes}
\alias{FazTexto.AccMes}
\title{Faz Texto Acumulado Mes}
\usage{
FazTexto.AccMes(x, digitos = 2, meses = 12)
}
\arguments{
\item{x}{Serie a ser calculada a taxa}

\item{digitos}{Numero de casas decimais. Por padrao, 2}

\item{meses}{Numero de meses para acumular. Por padrao, 12 meses}
}
\value{
Texto
}
\description{
Gera texto com taxa acumulada em m meses
}
\author{
Arthur Welle
}
