% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FazTexto.Cortes1.R
\name{FazTexto.Cortes1}
\alias{FazTexto.Cortes1}
\title{Faz Texto Cortes 1}
\usage{
FazTexto.Cortes1(x, z, y)
}
\arguments{
\item{x}{Serie a ser analisada}

\item{z}{Parâmetro a ser comparado}

\item{y}{Se 1, retorna uma palavra feminina. Se 2, retorna masculino}
}
\description{
Cria uma função que retorna texto “positivo” “negativo” dependendo se acima ou abaixo de z
}
\author{
Arthur Welle
}
