% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FazTexto.Cortes3.R
\name{FazTexto.Cortes3}
\alias{FazTexto.Cortes3}
\title{Faz Texto Corte 2}
\usage{
FazTexto.Cortes3(x, y, z, w)
}
\arguments{
\item{x}{Serie a ser analisada}

\item{y}{Criterio para ruim}

\item{z}{Criterio para bom}

\item{w}{Criterio para muito bom}
}
\value{
Texto
}
\description{
Avalia se a serie é muito ruiom, ruim, boa, muito boa.
}
\author{
Arthur Welle
}
