% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FazTexto.N_Indice.R
\name{FazTexto.N_Indice}
\alias{FazTexto.N_Indice}
\title{FazTexto.N_Indice}
\usage{
FazTexto.N_Indice(x)
}
\arguments{
\item{x}{Serie temporal}
}
\value{
Numero indice
}
\description{
Cria funcao para criar numero indice de uma serie mensal de inflação (para usar em deflacionamentos até para a ultima data)
}
\details{
Corrigir: Alterar nome
}
\author{
Arthur Welle
}
\keyword{Texto}
