% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FazTexto.TaxaAccMeses.R
\name{FazTexto.TaxaAccMeses}
\alias{FazTexto.TaxaAccMeses}
\title{Faz Texto Taxa Acumulada em M meses}
\usage{
FazTexto.TaxaAccMeses(x, meses = 12)
}
\arguments{
\item{x}{Serie a ser calculada a taxa acumulada}

\item{meses}{Numero de meses para acumular. Por padrao, 12 meses}
}
\value{
Retorna texto da taxa acumulada em m meses
}
\description{
Calcula taxa acumulada em m meses
}
\author{
Arthur Welle
}
