% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FazTexto.TaxaAnualizada.R
\name{FazTexto.TaxaAnualizada}
\alias{FazTexto.TaxaAnualizada}
\title{Faz Texto Taxa Anualizada}
\usage{
FazTexto.TaxaAnualizada(x, digitos = 2, texto = TRUE, tempo = 12)
}
\arguments{
\item{x}{Serie para calcular taxa anualizada}

\item{digitos}{Numero de casas decimais}

\item{texto}{Se TRUE retorna caractere}

\item{tempo}{Se mensal 12, trimestral 4. Se for xts, ts ou zoo, calcula automaticamente.}
}
\description{
Cria texto para taxa anualizada
}
\author{
Arthur Welle
}
