% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FazTexto.UltimoValor.R
\name{FazTexto.UltimoValor}
\alias{FazTexto.UltimoValor}
\title{Faz Texto Ultimo Valor}
\usage{
FazTexto.UltimoValor(x, digitos = 2)
}
\arguments{
\item{x}{Serie a ser analisada}

\item{digitos}{Numero de casas decimais. Por padrao, 2}
}
\value{
Texto
}
\description{
Cria uma funcao que retorna o ultimo valor de uma serie x.
}
\author{
Arthur Welle
}
