% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FazTexto.ValorMesAntes.R
\name{FazTexto.ValorMesAntes}
\alias{FazTexto.ValorMesAntes}
\title{Faz Texto Valor Meses Antes}
\usage{
FazTexto.ValorMesAntes(x, digitos = 2, meses = 12)
}
\arguments{
\item{x}{Serie a ser analisada}

\item{digitos}{Numero de casas decimais. Por padrao, 2}

\item{meses}{Numero de meses a serem analisados}
}
\value{
Texto
}
\description{
Retorna Valor de m meses anteriores
}
\author{
Arthur Welle
}
