% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FazTexto.VarMes.Abs.R
\name{FazTexto.VarMes.Abs}
\alias{FazTexto.VarMes.Abs}
\title{Faz Texto Variacao M meses absoluta}
\usage{
FazTexto.VarMes.Abs(x, digitos = 2, meses = 12, texto = TRUE)
}
\arguments{
\item{x}{Serie percentual}

\item{digitos}{Numero de casas decimais. Por padrao, 2}

\item{meses}{Numero de mesesa serem calculados.Por padrao, 12}

\item{texto}{Se TRUE, retorna valor em caractere}
}
\value{
Texto
}
\description{
Cria uma funcao que retorna a variação absoluta (em pontos percentuais se ja for uma serie percental) de uma serie no ultimo ano
}
\author{
Arthur Welle
}
