% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FazTexto.VarMes.porc.R
\name{FazTexto.VarMes.porc}
\alias{FazTexto.VarMes.porc}
\title{Faz Texto Variacao m meses percentual}
\usage{
FazTexto.VarMes.porc(x, digitos = 2, meses = 12, porcentagem = FALSE,
  texto = TRUE)
}
\arguments{
\item{x}{Serie a ser analisada}

\item{digitos}{Numero de casas decimias. Por padrao, 2}

\item{meses}{Numero de meses a serem calculados. Por padrao, 12}

\item{porcentagem}{se TRUE, incui % ao numero}

\item{texto}{Se TRUE retorna caractere}
}
\value{
Texto
}
\description{
Cria uma funcao que retorna a variacao percentuais de uma serie nos ultimos m meses
}
\author{
Arthur Welle
}
