% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_cecon.R
\name{grafico_cecon}
\alias{grafico_cecon}
\title{Grafico CECON}
\usage{
grafico_cecon(dado_xts, logo = TRUE, FUN = grafico_padrao, ...)
}
\arguments{
\item{dado_xts}{Dado em xts}

\item{logo}{Se TRUE, inclui logo do CECON}

\item{FUN}{Tipo de função que gera gráfico. Padrao grafico_padrao}

\item{...}{Argumentos adicionais para a funcao de grafico}
}
\value{
Grafico interativo em html
}
\description{
Gera grafico interatico com logo do cecon
}
\author{
Gabriel Petrini
}
