% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_padrao.R
\name{grafico_padrao}
\alias{grafico_padrao}
\title{Gráfico Padrão}
\usage{
grafico_padrao(dado_xts, tipo_grafico = geom_line(size = 1),
  titulo = NULL, fonte = NULL, x_titulo = NULL, y_titulo = NULL,
  tema = theme_classic(), quebra_data = "1 year", label_data = "\%Y",
  pontos = 0)
}
\arguments{
\item{dado_xts}{Objeto em xts ou que possa ser convertido para xts pela função ajuste_xts()}

\item{tipo_grafico}{Tipo de gráfico a ser plotado. Por padrão, é plotado um gráfico de linha}

\item{titulo}{Caractere que especifica qual o título do gráfico}

\item{fonte}{Caractere que espeficia a fonte. Não é necessário escrever "Fonte: "}

\item{x_titulo}{Caractere com título do eixo x}

\item{y_titulo}{Caractere com título do eixo y}

\item{tema}{Tema do gráfico. Por padrão, utiliza-se theme_classic()}

\item{quebra_data}{Período para seccionar o eixo x. Por padrão, é dividido por ano, isto é, "1 year"}

\item{label_data}{Formato em que é disponibilizada as datas do eixo x. Por padrão, é apresentada em anos, isto é, "%Y"}

\item{pontos}{Argumento para especificar se devem conter pontos no gráfico. Por padrão, não apresenta pontos, isto é, pontos = 0}
}
\value{
Retorna um gráfico ggplot
}
\description{
Gera um gráfico com um template básico
}
\author{
Gabriel Petrini
}
