% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_rapido.R
\name{grafico_rapido}
\alias{grafico_rapido}
\title{Grafico Rapido}
\usage{
grafico_rapido(dado, remover_NA = FALSE)
}
\arguments{
\item{dado}{Serie conversivel em xts}

\item{remover_NA}{Se TRUE, remove NAs}
}
\value{
Grafico interativo com logo do CECON
}
\description{
Gera um grafico interativo com logo do CECON a partir de uma serie conversivel em xts
}
\author{
Gabriel Petrini
}
