% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_movel.R
\name{media_movel}
\alias{media_movel}
\title{Media Movel}
\usage{
media_movel(dados, digitos = 2, meses = 12, tabela = FALSE)
}
\arguments{
\item{dados}{Serie conversivel para xts a ser calcular media movel}

\item{digitos}{numero de casas decimais}

\item{meses}{Numero de meses para calcular media movel}

\item{tabela}{Se TRUE, agrega os dados em uma tabela}
}
\value{
Media movel em xts
}
\description{
Calcula media movel
}
\author{
Arthur Welle
}
