% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salvar_grafico.R
\name{salvar_grafico}
\alias{salvar_grafico}
\title{Salvar Grafico}
\usage{
salvar_grafico(grafico, pasta = "./Graficos", formato = ".png")
}
\arguments{
\item{grafico}{Grafico gerado pelo ggplot}

\item{pasta}{Caminha onde salvar o grafico}

\item{formato}{Formato a ser salvo}
}
\value{
Imagem na pasta especificada
}
\description{
Salvar objeto ggplot em uma pasta especifica
}
\author{
Gabriel Petrini
}
