% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_acum.R
\name{tx_acum}
\alias{tx_acum}
\title{Taxa Acumulada}
\usage{
tx_acum(dados, digitos = 2, meses = 12, tabela = FALSE)
}
\arguments{
\item{dados}{Objeto conversivel em xts}

\item{digitos}{Numero de casas decimais. Por padrao, 2}

\item{meses}{Numero de meses para acumular a taxa de crescimento}

\item{tabela}{Se TRUE, retorna a taxa acumulada junto com os dados originais}
}
\value{
Um objeto xts
}
\description{
Retorna taxa acumulada em m meses
}
\author{
Arthur Welle
}
\keyword{Transformacoes}
