% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FazTexto.TaxaAnualizada.R
\name{FazTexto.TaxaAnualizada}
\alias{FazTexto.TaxaAnualizada}
\title{Faz Texto Taxa Anualizada}
\usage{
FazTexto.TaxaAnualizada(x, digitos = 2)
}
\arguments{
\item{x}{Serie para calcular taxa anualizada}

\item{digitos}{Numero de casas decimais}
}
\description{
Cria texto para taxa anualizada
}
\author{
Arthur Welle
}
