% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_padrao.R
\name{grafico_padrao}
\alias{grafico_padrao}
\title{Grafico Padrao}
\usage{
grafico_padrao(dado_xts, tipo_grafico = geom_line(size = 1),
  titulo = NULL, fonte = NULL, x_titulo = NULL, y_titulo = NULL,
  tema = theme_classic(), quebra_data = "1 year", label_data = "\%Y",
  pontos = 0)
}
\arguments{
\item{dado_xts}{Objeto em xts ou que possa ser convertido para xts pela funcao ajuste_xts()}

\item{tipo_grafico}{Tipo de grafico a ser plotado. Por padrao, grafico de linha}

\item{titulo}{Caractere que especifica qual o titulo do grafico}

\item{fonte}{Caractere que espeficia a fonte. Nao e necessario escrever "Fonte: "}

\item{x_titulo}{Caractere com titulo do eixo x}

\item{y_titulo}{Caractere com titulo do eixo y}

\item{tema}{Tema do grafico. Por padrao, utiliza-se theme_classic()}

\item{quebra_data}{Periodo para seccionar o eixo x. Por padrao, e dividido por ano, isto e, "1 year"}

\item{label_data}{Formato em que e disponibilizada as datas do eixo x. Por padrao, e apresentada em anos, isto e, "%Y"}

\item{pontos}{Argumento para especificar se devem conter pontos no grafico. Por padrao, nao apresenta pontos, isto e, pontos = 0}
}
\value{
Retorna um grafico ggplot
}
\description{
Gera um grafico com um template basico
}
\author{
Gabriel Petrini
}
